refactoring: extracted DiagnosticImpl to separate class
